+++ src/full/Agda/Syntax/DoNotation.hs
@@ -62,8 +62,12 @@ desugarDo qBind qThen [] = genericError "Empty 'do' block"
 
 -- The last statement must be a DoThen
 desugarDo qBind qThen [s]
-  | DoThen e <- s = return e
-  | otherwise     = genericError "The last statement in a 'do' block must be an expression"
+  | DoThen e              <- s           = return e
+  | DoBind r p e []       <- s
+  , Just (r' , NotHidden) <- isAbsurdP p =
+      return $ appOp (setRange r qBind) e $ AbsurdLam r' NotHidden
+  | otherwise                            =
+      genericError "The last statement in a 'do' block must be an expression or an absurd match."
 
 -- `DoThen` and `DoLet` are easy
 desugarDo qBind qThen (DoThen e   : ss) = appOp qThen e   <$> desugarDo qBind qThen ss